+++ Sources/StreamVideo/WebRTC/Screensharing/BroadcastBufferReaderConnection.swift
@@ -5,19 +5,14 @@
 import Foundation
 import Darwin
 
-final class BroadcastBufferReaderConnection: NSObject {
+final class BroadcastBufferReaderConnection: BroadcastBufferConnection {
     private let streamDelegate: StreamDelegate
     
     private let filePath: String
     private var socketHandle: Int32 = -1
     private var address: sockaddr_un?
     
-    private var inputStream: InputStream?
-    private var outputStream: OutputStream?
-    
     private var listeningSource: DispatchSourceRead?
-    private var streamQueue: DispatchQueue?
-    private var shouldKeepRunning = false
     
     init?(filePath path: String, streamDelegate: StreamDelegate) {
         self.streamDelegate = streamDelegate
@@ -47,7 +42,11 @@ final class BroadcastBufferReaderConnection: NSObject {
                 return
             }
             
-            self.setupStreams(clientSocket: clientSocket)
+            self.setupStreams(
+                clientSocket: clientSocket,
+                delegate: streamDelegate,
+                handleOutput: false
+            )
             self.openStreams()
         }
         
@@ -56,34 +55,13 @@ final class BroadcastBufferReaderConnection: NSObject {
         return true
     }
     
-    func close() {
-        unscheduleStreams()
-        closeStreams()
+    override func close() {
+        super.close()
         listeningSource?.cancel()
         Darwin.close(socketHandle)
     }
     
-    func writeToStream(buffer: UnsafePointer<UInt8>, maxLength length: Int) -> Int {
-        outputStream?.write(buffer, maxLength: length) ?? 0
-    }
-    
     //MARK: - private
-    
-    private func openStreams() {
-        inputStream?.open()
-        outputStream?.open()
-    }
-    
-    private func closeStreams() {
-        inputStream?.delegate = nil
-        outputStream?.delegate = nil
-        
-        inputStream?.close()
-        outputStream?.close()
-        
-        inputStream = nil
-        outputStream = nil
-    }
     
     private func setupAddress() -> Bool {
         var addr = sockaddr_un()
@@ -122,47 +100,5 @@ final class BroadcastBufferReaderConnection: NSObject {
         }
         
         return true
-    }
-    
-    private func setupStreams(clientSocket: Int32) {
-        var readStream: Unmanaged<CFReadStream>?
-        var writeStream: Unmanaged<CFWriteStream>?
-        
-        CFStreamCreatePairWithSocket(kCFAllocatorDefault, clientSocket, &readStream, &writeStream)
-        
-        inputStream = readStream?.takeRetainedValue()
-        inputStream?.delegate = self.streamDelegate
-        inputStream?.setProperty(kCFBooleanTrue, forKey: Stream.PropertyKey(kCFStreamPropertyShouldCloseNativeSocket as String))
-        
-        outputStream = writeStream?.takeRetainedValue()
-        outputStream?.setProperty(kCFBooleanTrue, forKey: Stream.PropertyKey(kCFStreamPropertyShouldCloseNativeSocket as String))
-        
-        scheduleStreams()
-    }
-    
-    private func scheduleStreams() {
-        shouldKeepRunning = true
-        
-        streamQueue = DispatchQueue.global(qos: .userInitiated)
-        streamQueue?.async { [weak self] in
-            self?.inputStream?.schedule(in: .current, forMode: .default)
-            self?.outputStream?.schedule(in: .current, forMode: .default)
-            
-            var isRunning = false
-            
-            repeat {
-                isRunning = self?.shouldKeepRunning ?? false
-                && RunLoop.current.run(mode: .default, before: .distantFuture)
-            } while (isRunning)
-        }
-    }
-    
-    private func unscheduleStreams() {
-        streamQueue?.sync { [weak self] in
-            self?.inputStream?.remove(from: .current, forMode: .common)
-            self?.outputStream?.remove(from: .current, forMode: .common)
-        }
-        
-        shouldKeepRunning = false
     }
 }
